refactor(mcp): split main.rs into tool_schemas and window_capture modules
Audit step 17: agentmux-mcp/src/main.rs was 4,727 lines. Two clean seams
come out, leaving 3,239.

- tool_schemas.rs (716): the 50 *_TOOL JSON schema consts. Pure data, and
  ~15% of the file — every schema tweak used to diff against the file that
  also holds the dispatcher. All 50 bodies are byte-identical to main
  (verified mechanically against git show) and the declaration order is
  unchanged, so the advertised tool list is the same.
- window_capture.rs (814): the largest single cluster in the file, and one
  concern — find AgentMux windows, decide what this agent may capture
  (CaptureTier), capture it, record it. Coupling was measured before the
  move: it calls exactly one outside function (agent_slug) and uses no
  outside types.

main.rs glob-imports both, so call_tool's call sites and mod tests's
'use super::*' resolve as before — the test module neither moved nor
changed. Visibility changes are the minimum the compiler demanded:
pub(crate) on the 19 items named from outside, their 11 struct fields,
CaptureTier::{allowed,label}, and agent_slug. Nothing became pub.

cargo test -p agentmux-mcp: 26/26. Full workspace: 4,072 passed, 0 failed.
