+++ .changesets/1788783575-refactor-mcp-split-agentmux-mcp-src-main-rs-50-tool-schemas--2bbe.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(mcp): split agentmux-mcp/src/main.rs — 50 tool schemas into tool_schemas.rs and the window capture/discovery cluster into window_capture.rs (audit step 17); no behavior change
