+++ agentmux-mcp/src/main.rs
@@ -55,706 +55,10 @@ struct LoopEntry {
 /// when the agent pane closes.
 type LoopRegistry = Mutex<HashMap<String, LoopEntry>>;
 
-const SHELL_TOOL: &str = r#"{
-  "name": "Shell",
-  "description": "Start a long-running shell process. Returns immediately with a shell_id. Output streams live in the conversation document. Use for build systems, watchers, dev servers — anything that should run in the background without blocking the conversation. Stop it later with ShellStop(shell_id) — never use kill/taskkill on a shell you started.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "cmd":            { "type": "string",  "description": "Command to run (passed to sh -c / cmd /C)" },
-      "cwd":            { "type": "string",  "description": "Working directory (defaults to agent workdir)" },
-      "title":          { "type": "string",  "description": "Display label shown in the conversation row (defaults to cmd)" },
-      "env":            { "type": "object",  "description": "Extra environment variables", "additionalProperties": { "type": "string" } },
-      "capture_stdin":  { "type": "boolean", "description": "Pipe stdin so ShellInput() can write to it. Default false — avoids blocking programs that read stdin to EOF (e.g. `cat` with no args). Set true only when you intend to use ShellInput()." }
-    },
-    "required": ["cmd"]
-  }
-}"#;
-
-const SHELL_STOP_TOOL: &str = r#"{
-  "name": "ShellStop",
-  "description": "Stop a running shell started by Shell(). Pass the shell_id returned by Shell. Tree-kills the whole process group (e.g. `task dev` and its child task.exe/node processes), so prefer this over kill/taskkill — those can hit other agents' or the host's processes by name.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "shell_id": { "type": "string", "description": "The shell_id returned by a prior Shell() call" }
-    },
-    "required": ["shell_id"]
-  }
-}"#;
-
-const SHELL_INPUT_TOOL: &str = r#"{
-  "name": "ShellInput",
-  "description": "Write text to the stdin of a running shell started by Shell(). A newline is appended automatically. Use for interactive prompts like 'Terminate batch job (Y/N)?' or REPL commands. REQUIRES the shell to have been created with capture_stdin=true; otherwise its stdin is /dev/null and this returns an error telling you to recreate the shell. Also returns an error if the shell has exited. Note: processes that block waiting for stdin-EOF (e.g. `cat` with no args) will not exit until ShellStop is called — ShellStop always unblocks them via kill.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "shell_id": { "type": "string", "description": "The shell_id returned by a prior Shell() call" },
-      "text":     { "type": "string", "description": "Text to write to stdin (newline appended automatically)" }
-    },
-    "required": ["shell_id", "text"]
-  }
-}"#;
-
-const SHELL_STATUS_TOOL: &str = r#"{
-  "name": "ShellStatus",
-  "description": "Query whether a shell started by Shell() is still running. Returns running status, exit code (when exited), and total line count so far. Use to poll for completion or check if a dev server is still up before starting a second one.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "shell_id": { "type": "string", "description": "The shell_id returned by a prior Shell() call" }
-    },
-    "required": ["shell_id"]
-  }
-}"#;
-
-const SEND_MESSAGE_TOOL: &str = r#"{
-  "name": "SendMessage",
-  "description": "Send a message to another agent by name. The message is injected as input into the target agent's active conversation. Use for agent-to-agent coordination — handoff, task delegation, status notifications. Delivery is best-effort and tries local → same-host → LAN → cloud in order. The first three tiers return only once the message has actually been injected; the cloud tier is store-and-forward, so success there means the relay accepted it and the recipient's AgentMux will pick it up when it next syncs (which never happens if that instance stays offline). Returns once one tier has taken the message or all have failed.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "to":      { "type": "string", "description": "Name of the target agent (its AGENTMUX_AGENT_ID value)" },
-      "message": { "type": "string", "description": "Message text to inject into the target agent's conversation" }
-    },
-    "required": ["to", "message"]
-  }
-}"#;
-
-const DISCOVER_AGENTS_TOOL: &str = r#"{
-  "name": "DiscoverAgents",
-  "description": "List the agents and instances reachable from here across the muxbus delivery tiers (host, LAN, cloud), so you can pick a valid target before SendMessage. Returns JSON: host.addressable (agents reachable right now via local delivery), host.agents (this host's agent directory, each with an `addressable` flag), lan (LAN peers and the agents on them), and wan.local_agents_subscribed (THIS host's own agents that are subscribed to the cloud relay — not a list of remote agents you can reach over WAN; the relay exposes no such directory, so a cloud-only target simply won't appear anywhere in this output). Addressing is by agent name, case-insensitive. Takes no arguments.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-const GET_AGENT_TRANSCRIPT_TOOL: &str = r#"{
-  "name": "GetAgentTranscript",
-  "description": "Read the tail of a registered agent's session transcript by name, plus whether it currently has a turn in flight (turn_active). Resolves agents on this host first, then falls back to other channels on this same host (cross-channel) — does NOT reach LAN or WAN agents (use ListConversations to see those, but note they're liveness-only for now). For a Warden Supervisor watcher agent polling other agents on its own interval to decide whether to nudge a stalled one to continue, or any agent wanting to check what another agent on this host is doing. Returns JSON: {agent, block_id, tier, turn_active, lines: [...], truncated}. Read-only, best-effort — does not deliver anything to the target.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "agent":     { "type": "string", "description": "Name of the target agent (its AGENTMUX_AGENT_ID value)" },
-      "max_lines": { "type": "integer", "description": "Max number of recent transcript lines to return (default 100, server-capped at 500)" }
-    },
-    "required": ["agent"]
-  }
-}"#;
-
-// Cross-tier conversation glance
-// (SPEC_MUXSPECT_CROSS_TIER_CONVERSATION_VISIBILITY_2026_08_21.md Phase A) —
-// one call to see every agent's recent activity across host + cross-channel
-// (with a last-message preview) and LAN + WAN (liveness only — Phase A
-// deliberately does not invent a remote-read protocol for those tiers; see
-// the spec's Phase B/C). Thin passthrough to the srv's own
-// `/api/v1/muxspect/conversations`, same auth pattern as every other tool
-// here.
-const LIST_CONVERSATIONS_TOOL: &str = r#"{
-  "name": "ListConversations",
-  "description": "See every agent's most recent activity across host, cross-channel (other AgentMux channels on this same host), LAN, and connected WAN in one call — a faster alternative to DiscoverAgents + N x GetAgentTranscript. Host and cross-channel entries include turn_active, last_activity_ms, and a last_message_preview (tail transcript line). LAN and WAN entries are liveness-only (remote_fetch_required: true) — reading their conversation content isn't supported yet. Read-only. Takes no arguments.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-const SUPERVISOR_NUDGE_TOOL: &str = r#"{
-  "name": "SupervisorNudge",
-  "description": "For a Warden Supervisor watcher agent: record a decision about a target agent that just ended its turn, after inspecting it with GetAgentTranscript. action=\"nudge\" delivers a fixed, server-owned continuation message to the target as an ordinary jekt and logs the decision — this tool does NOT accept custom message text; the nudge is deliberately a narrow, non-composable template, not an instruction you write per-situation. action=\"decline\" sends nothing and just logs that you chose not to nudge (e.g. the target isn't opted in, or looks genuinely done/blocked, not merely pausing to ask). A nudge is refused (tool error) if the target hasn't opted in via auto_continue_enabled, or if it would exceed the server-side consecutive-nudge ceiling — treat either as a signal to stop nudging this target and escalate to a human via SendMessage instead of retrying.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "target_agent": { "type": "string", "description": "Name of the agent this decision is about (its AGENTMUX_AGENT_ID value)" },
-      "action":       { "type": "string", "enum": ["nudge", "decline"], "description": "Whether to nudge the target to continue or decline to" },
-      "reason":       { "type": "string", "description": "Your stated reasoning for this decision, recorded in the audit log" }
-    },
-    "required": ["target_agent", "action"]
-  }
-}"#;
-
-const SET_ACTIVE_TAB_TOOL: &str = r#"{
-  "name": "SetActiveTab",
-  "description": "Switch the active (foreground) tab to the given tab_id within its workspace. Get tab ids from Layout(query:\"tabs\") or Layout(query:\"layout\").",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "tab_id": { "type": "string", "description": "The tab to make active" }
-    },
-    "required": ["tab_id"]
-  }
-}"#;
-
-const NEW_TAB_TOOL: &str = r#"{
-  "name": "NewTab",
-  "description": "Open a new tab in your own workspace and switch to it. Optionally name it; otherwise AgentMux auto-names it.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "name": { "type": "string", "description": "Optional name for the new tab" }
-    }
-  }
-}"#;
-
-const FOCUS_WINDOW_TOOL: &str = r#"{
-  "name": "FocusWindow",
-  "description": "Bring an AgentMux window to the foreground. Defaults to your own window. Pass window_id (from Layout(query:\"windows\") or Layout(query:\"layout\")) to focus a specific one.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "window_id": { "type": "string", "description": "Optional window to focus; defaults to your own" }
-    }
-  }
-}"#;
-
-const UI_SCREENSHOT_TOOL: &str = r#"{
-  "name": "UIScreenshot",
-  "description": "Capture a screenshot of your OWN AgentMux pane's UI — clipped to just that pane, not the whole shared window (can't see other panes/agents). Returns a file path; use the Read tool on that path to view the image yourself, or OpenMedia to show it to the user. Use this to visually verify a UI change (e.g. after UIClick-ing a button).",
-  "inputSchema": { "type": "object", "properties": {} }
-}"#;
-
-const UI_CLICK_TOOL: &str = r#"{
-  "name": "UIClick",
-  "description": "Click an element in AgentMux's UI — a real synthesized mouse click (not a scripted .click()), so focus/hover/pointer behavior matches a human click. Reaches your OWN pane and shared app chrome (status bar, hamburger menu, window controls); cannot reach a DIFFERENT pane or agent's UI. Use UIQuery first if you're not sure of the right CSS selector.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "selector": { "type": "string", "description": "CSS selector for the element to click, scoped to your own pane" }
-    },
-    "required": ["selector"]
-  }
-}"#;
-
-const UI_QUERY_TOOL: &str = r#"{
-  "name": "UIQuery",
-  "description": "Find elements in AgentMux's UI matching a CSS selector — returns tag, text, attributes, bounding rect, and focus state for each match. Reaches your OWN pane and shared app chrome (status bar, hamburger menu, window controls); cannot reach a DIFFERENT pane or agent's UI. Use this to locate an element before UIClick-ing it, or to read rendered text/state (e.g. did a button's label change) without taking a screenshot.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "selector": { "type": "string", "description": "CSS selector to match, scoped to your own pane" },
-      "limit": { "type": "number", "description": "Max number of matches to return (default: all)" }
-    },
-    "required": ["selector"]
-  }
-}"#;
-
-// Deliberately NOT part of the ui_handlers.rs signed-identity/pane-ownership
-// scheme UIScreenshot/UIClick/UIQuery use. Authorization here is by
-// `CaptureTier` (what is being captured), not by pane ownership.
-//
-// **This block previously asserted three invariants that are no longer true**
-// (reagentx P2 on PR #2845 — the same stale-comment class already fixed on
-// `enumerate_agentmux_windows`, missed here on the first pass):
-//   - "crosses no agent-to-agent trust boundary within one instance" — it
-//     does now. T1 capture reaches this instance's own window, which contains
-//     every agent's pane.
-//   - "the caller's OWN instance is always excluded from the candidate set" —
-//     no longer excluded; that is the point of the change.
-//   - "Scoped to AgentMux's own windows only" — non-AgentMux windows are T4
-//     targets now.
-//
-// What replaced them, per
-// `SPEC_AGENT_UNRESTRICTED_CAPTURE_WITH_ACCOUNTABILITY_2026_08_30.md`: the
-// repo owner directed that agents be able to capture anything, so the control
-// moved from prevention to accountability. Every agent-to-agent tier is open
-// and audited; only T3 (a window owned by a DIFFERENT OS user) is withheld,
-// because that is a human boundary rather than an agent one — that person
-// never consented and cannot be notified.
-//
-// Two things reagent P1 (PR #2709 round 2) established still hold and must
-// not be undone while widening scope: the `is_agentmux` flag still gates
-// DISCLOSURE even though it no longer gates inclusion — foreign windows stay
-// out of `DiscoverWindows`' default listing and their titles are withheld
-// from candidate/miss lists (`candidate_label`). That is what keeps round 2's
-// KeePass-title leak closed now that foreign windows are enumerated at all.
-//
-// The per-app approval gating, mutex and app-tier warnings that
-// docs/specs/computer-use-pane.md (draft, unbuilt) scopes for arbitrary
-// third-party app control remain OUT of scope here: this is read-only pixel
-// capture, not input injection.
-const CAPTURE_WINDOW_TOOL: &str = r#"{
-  "name": "CaptureWindow",
-  "description": "Screenshot any window on this machine — another AgentMux instance, your OWN instance's other windows (including a torn-off floating pane), or a non-AgentMux application. Target by pid (preferred; get one from DiscoverWindows) or by (partial, case-insensitive) title match. Prefer pid: AgentMux's own frontend actively rewrites window titles (tab switches, workspace renames), so a title match can go stale mid-session in a way a pid never does. The ONE exception is a window owned by a DIFFERENT OS user — withheld by default, because that is a human boundary rather than an agent one and no in-app notification can reach that person. If title_contains matches more than one window and you did not pass an explicit index, this returns every candidate (pid + title) instead of guessing — pass one of those pids back in, or use DiscoverWindows first. A single process can also own multiple top-level windows sharing the same pid — if pid matches more than one, pass an explicit index alongside it the same way. Every call is logged (who, what, which tier, a hash of the image, outcome) to an audit trail in this instance's own data dir. Returns a file path; use the Read tool on that path to view it yourself, or OpenMedia to show it to the user. The result may note likely_unrendered if the captured frame looks solid/near-solid-color even after a couple of internal retries — that CAN mean the window has not painted its first real frame yet, but it is only a heuristic (a legitimately solid-colored window trips it too).",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "pid": { "type": "number", "description": "Process id of the target window's owning process — from DiscoverWindows. Preferred over title_contains: stable for the process's whole lifetime, unlike its title. If more than one window shares this pid, also pass index to disambiguate." },
-      "title_contains": { "type": "string", "description": "Substring to match against capturable windows' titles, case-insensitive — that includes non-AgentMux windows, not just AgentMux ones. Ignored if pid is given. Note the candidate/miss lists returned on an ambiguous or failed match identify a non-AgentMux window by pid only, never by title." },
-      "index": { "type": "number", "description": "If multiple windows match title_contains (or share the given pid), which one to capture (0-based). Omit this when you expect exactly one match — if more than one actually matches, the tool returns the full candidate list instead of silently picking one." }
-    }
-  }
-}"#;
-
-const DISCOVER_WINDOWS_TOOL: &str = r#"{
-  "name": "DiscoverWindows",
-  "description": "List top-level windows on this machine — read-only: no screenshot taken, nothing written to disk. Use this BEFORE CaptureWindow so you have real candidates (pid, title, exe_path) instead of guessing a title substring. Each entry reports its capture `tier` and whether it is `capturable`. A window owned by a different OS user is never capturable; when it IS listed it appears withheld rather than omitted — `capturable: false`, `title` and `exe_path` null, plus a `withheld_reason` — so you can see that it exists and why it is out of reach without its content crossing that boundary. Whether it is listed at all follows the same `include_foreign` gate as any other window: by default only AgentMux windows are listed, so another user's non-AgentMux window shows up only with `include_foreign: true`. By default lists AgentMux windows only; pass include_foreign to also list other applications (kept opt-in so ordinary discovery does not disclose the titles of a user's unrelated apps). exe_path can reveal the OS username of a different instance's owner on a shared machine, so — same as CaptureWindow — every call is logged to this instance's own audit trail.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "include_self": { "type": "boolean", "description": "Include this agent's own AgentMux instance's window(s) in the results. Default false, since they are usually not what you are looking for — but unlike before, they ARE capturable, so pass true when you want your own instance's other windows (e.g. a torn-off floating pane)." },
-      "include_foreign": { "type": "boolean", "description": "Also list non-AgentMux windows (other applications). Default false so ordinary discovery does not disclose the titles of a user's unrelated apps as a side effect. These are capturable when listed." }
-    }
-  }
-}"#;
-
-// Fleet control (SPEC_MULTI_AGENT_FLEET_CONTROL_2026_08_20.md) — select,
-// broadcast, and bulk-act on many agents at once, from an agent's own
-// perspective. FleetList is a thin, fleet-framed alias of DiscoverAgents
-// (same `/agentmux/discovery` call — its response already carries each
-// reachable agent's `block_id`, exactly what FleetBroadcast/FleetBulkStop
-// need as `targets`). FleetBroadcast loops the SAME signed single-target
-// delivery path SendMessage uses, once per target, entirely client-side —
-// only this process holds AGENTMUX_JEKT_KEY, so per-message signing can
-// only happen here, never in a server-side batch RPC. FleetBulkStop calls
-// the one fleet action that genuinely IS a server-side batch RPC
-// (`POST /api/v1/fleet/bulk-stop`) since stopping a controller involves no
-// jekt signing at all.
-const FLEET_LIST_TOOL: &str = r#"{
-  "name": "FleetList",
-  "description": "List every agent reachable from here (same data as DiscoverAgents), framed for fleet targeting: each entry's block_id is what FleetBroadcast/FleetBulkStop expect in their `targets` array. Use this before either of those to build your target list. Read-only. Takes no arguments.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-const FLEET_BROADCAST_TOOL: &str = r#"{
-  "name": "FleetBroadcast",
-  "description": "Send the SAME message to many agents at once (get targets from FleetList). Delivers each one individually and signed, exactly like SendMessage would — this is a convenience loop, not a new delivery mechanism. Targets reach every tier FleetList/DiscoverAgents can see: a host or cross-channel target is its block_id; a LAN or WAN target (no block_id exists for those — they're not local blocks) is its agent NAME instead — pass whichever FleetList gave you for that entry. Returns JSON {succeeded: [target...], failed: [{id, error}...]} — always check `failed`, a partial failure is common (e.g. one target went offline) and is never silently dropped.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "targets": { "type": "array", "items": { "type": "string" }, "description": "block_id (host/cross-channel) or agent name (LAN/WAN) values to send to — from FleetList" },
-      "message": { "type": "string", "description": "Message text to inject into each target's conversation" }
-    },
-    "required": ["targets", "message"]
-  }
-}"#;
-
-const OPEN_AGENT_TOOL: &str = r#"{
-  "name": "OpenAgent",
-  "description": "Launch an agent into a pane by name or definition id (get these from FleetList/DiscoverAgents). Idempotent: if the agent is already open in the target tab, returns its existing pane with created:false instead of opening a second one. This spawns a real provider process (and, for container agents, execs into their container) — it consumes provider tokens like any human-opened pane. Returns JSON {block_id, tab_id, agent_id, provider, controller_type, status, created}. After a successful open the agent becomes addressable for SendMessage.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "agent_id": { "type": "string", "description": "Agent definition id or name (case-insensitive), e.g. from FleetList's host.agents[]" },
-      "tab_id": { "type": "string", "description": "Optional target tab id; defaults to the active tab" },
-      "focus": { "type": "boolean", "description": "Optional: focus the new pane (default true)" }
-    },
-    "required": ["agent_id"]
-  }
-}"#;
-
-const FLEET_BULK_STOP_TOOL: &str = r#"{
-  "name": "FleetBulkStop",
-  "description": "Stop many agent panes at once by block_id (get these from FleetList). Destructive — double-check your target list first. Returns JSON {succeeded, failed: [{id, error}...], aborted_early}. Optionally pass `staged` to cap blast radius on a bad selection: stops `batch_size` targets at a time, and if a batch's failure rate exceeds `max_fail_percentage`, the remaining targets are recorded as failed (untried) instead of being attempted — `aborted_early` will be true. Without `staged`, every target is attempted as one batch.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "targets": { "type": "array", "items": { "type": "string" }, "description": "block_id values to stop (from FleetList)" },
-      "signal": { "type": "string", "description": "Optional: SIGKILL or SIGTERM for a forceful stop; default is a graceful stop" },
-      "staged": {
-        "type": "object",
-        "description": "Optional staged rollout to cap blast radius",
-        "properties": {
-          "batch_size": { "type": "integer", "description": "How many targets to stop per batch" },
-          "max_fail_percentage": { "type": "integer", "description": "Abort remaining batches if a completed batch's failure rate exceeds this (0-100)" }
-        },
-        "required": ["batch_size", "max_fail_percentage"]
-      }
-    },
-    "required": ["targets"]
-  }
-}"#;
-
-// Consolidated read/introspection verb — replaces the former GetLayout /
-// ListWindows / ListWorkspaces / ListTabs tools (one tool, `query` selects the
-// view). `WhoAmI` stays its own no-arg tool (the spec's "foundation" self-context
-// call). See SPEC_AGENT_API_FIRST_CLASS_SURFACE_2026_06_17.md §4.6 / §10.
-const LAYOUT_TOOL: &str = r#"{
-  "name": "Layout",
-  "description": "Read the AgentMux UI structure around you. `query` selects what to return: \"layout\" (the full tree — every window with its workspace, tabs, and panes [block_id, view, title], and which tab is active), \"windows\" (window_id, display name, assigned workspace), \"workspaces\" (workspace_id, name, tab count, active tab), or \"tabs\" (tabs in your own workspace: tab_id, name, pane count). Read-only; use before naming or focusing things. For your OWN ids (block/tab/window/workspace), use WhoAmI instead.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "query": { "type": "string", "enum": ["layout", "windows", "workspaces", "tabs"], "description": "What to return (default: layout)" }
-    }
-  }
-}"#;
-
-const WHOAMI_TOOL: &str = r#"{
-  "name": "WhoAmI",
-  "description": "Return your own place in the AgentMux UI: your block (pane), tab, window, and workspace ids plus their names. Use it to discover the targets for naming/layout verbs (e.g. before SetName). Takes no arguments.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-// Consolidated naming verb — replaces the former SetWindowName / SetTabName /
-// SetPaneTitle / SetWorkspaceName tools (one tool, `target` selects which UI
-// element). All default to the caller's own element and are non-destructive.
-// See SPEC_AGENT_API_FIRST_CLASS_SURFACE_2026_06_17.md §4.3 / §10.
-const SET_NAME_TOOL: &str = r#"{
-  "name": "SetName",
-  "description": "Rename an AgentMux UI element. `target` selects which: \"window\" (the OS taskbar / window-title name; clamped to 64 chars), \"tab\" (the tab-bar label), \"pane\" (a conversation pane's header title), or \"workspace\" (the workspace name). Defaults to your own element; pass `target_id` (from Layout/WhoAmI) to rename any specific element by id. Names are trimmed; non-window names clamp to 128 chars.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "target":    { "type": "string", "enum": ["window", "tab", "pane", "workspace"], "description": "Which UI element to rename" },
-      "name":      { "type": "string", "description": "The new name/title to display" },
-      "target_id": { "type": "string", "description": "Explicit id of the element to rename (window_id / tab_id / workspace_id / block_id depending on target). Omit to default to your own." }
-    },
-    "required": ["target", "name"]
-  }
-}"#;
-
-const OPEN_EDITOR_TOOL: &str = r#"{
-  "name": "OpenEditor",
-  "description": "Open a file in an AgentMux editor pane next to this conversation. Use when you want the user to see a file you're discussing or editing. Pass an absolute host path. Fire-and-forget: returns once the pane is opened.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "file":  { "type": "string", "description": "Absolute path to the file to open" },
-      "title": { "type": "string", "description": "Optional tab/pane title (defaults to the file name)" },
-      "split": { "type": "string", "enum": ["right", "left", "down", "up"], "description": "Where to place the new pane relative to this agent pane (default: right). Ignored when floating is true." },
-      "collapse_tree": { "type": "boolean", "description": "Open the editor with its file-tree sidebar collapsed (just the file, no explorer). Default: false (tree expanded)." },
-      "floating": { "type": "boolean", "description": "Open the file in a floating window (a chromeless pane over the app) instead of a docked split. Default: false." }
-    },
-    "required": ["file"]
-  }
-}"#;
-
-const OPEN_MEDIA_TOOL: &str = r#"{
-  "name": "OpenMedia",
-  "description": "Open an image, video, or audio file in an AgentMux media pane next to this conversation. Use when you want the user to see/watch generated media you're discussing. Pass an absolute host path. Fire-and-forget: returns once the pane is opened.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "file":     { "type": "string", "description": "Absolute path to the media file to open" },
-      "title":    { "type": "string", "description": "Optional tab/pane title (defaults to the file name)" },
-      "split":    { "type": "string", "enum": ["right", "left", "down", "up"], "description": "Where to place the new pane relative to this agent pane (default: right). Ignored when floating is true." },
-      "floating": { "type": "boolean", "description": "Open the file in a floating window (a chromeless pane over the app) instead of a docked split. Default: false." }
-    },
-    "required": ["file"]
-  }
-}"#;
-
-const LOOP_TOOL: &str = r#"{
-  "name": "Loop",
-  "description": "Cross-agent recurring inject: run a prompt or slash command on a recurring interval by injecting it into ANOTHER agent's conversation (or your own, if you explicitly need muxbus-delivered self-messaging). Returns immediately with a loop_id; the prompt is injected on a fixed schedule until you call LoopStop(loop_id) or it exhausts max_iterations. If you're scheduling your OWN future turn (a same-session self-check, no other agent involved) — prefer the native ScheduleWakeup (one-off or adaptive-delay recurring, via re-arming) or native CronCreate (durable, cron-expression) tools instead: they have zero delivery overhead (no cross-agent messaging envelope) and built-in cache-window-aware backoff guidance this tool doesn't have. Use THIS tool only when the target is a different agent, or you specifically need AgentMux-persisted delivery across a restart. Loops stop automatically when the agent pane closes.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "prompt":         { "type": "string",  "description": "The prompt or slash command to inject each interval (e.g. 'check the PR status' or '/babysit-prs')" },
-      "interval":       { "type": "string",  "description": "How often to run: a number with optional unit s/m/h (e.g. '30s', '5m', '1h'). A bare number is minutes. Default '10m'. Minimum 10s." },
-      "to":             { "type": "string",  "description": "Target agent name (its AGENTMUX_AGENT_ID) to inject into. Defaults to this agent itself (a self-loop)." },
-      "immediate":      { "type": "boolean", "description": "Run once immediately on start in addition to every interval. Default false (first run after one interval)." },
-      "max_iterations": { "type": "integer", "description": "Stop automatically after this many fires. Omit or set to 0 for unlimited." }
-    },
-    "required": ["prompt"]
-  }
-}"#;
-
-const LOOP_STOP_TOOL: &str = r#"{
-  "name": "LoopStop",
-  "description": "Stop a recurring loop started by Loop(). Pass the loop_id it returned. Loops also stop automatically when the agent pane closes or max_iterations is reached.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "loop_id": { "type": "string", "description": "The loop_id returned by a prior Loop() call" }
-    },
-    "required": ["loop_id"]
-  }
-}"#;
-
-const LOOP_LIST_TOOL: &str = r#"{
-  "name": "LoopList",
-  "description": "List all currently running loops in this agent session. Returns each loop's id, prompt, target, interval, fire count, and remaining iterations (if capped). Like 'ps' for loops.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-// ── Muxqueue — the universal agent work queue ───────────────────────────────
-// docs/reports/REPORT_UNIVERSAL_AGENT_WORK_QUEUE_2026_09_01.md
-//
-// The pull/unaddressed/deferred counterpart to SendMessage's push/addressed/
-// immediate delivery: work goes in without naming a recipient, and whichever
-// agent asks next takes it. Deliberately NOT modelled on Cron (time-triggered)
-// or Loop (repeating) — the trigger here is an agent being ready.
-
-const WORK_ENQUEUE_TOOL: &str = r#"{
-  "name": "WorkEnqueue",
-  "description": "Put a unit of work on the shared Muxqueue for ANY agent to pick up later. Use this instead of SendMessage when you do NOT need a specific agent, or need it done eventually rather than now — 'someone should repro this', 'this PR needs review when a reviewer frees up'. The item persists across pane closes, app restarts, and version/channel changes, and is visible to every agent on this machine. If you know exactly who should do it and it should happen immediately, use SendMessage instead; if it should happen on a schedule, use CronCreate. Returns the item id.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "title":        { "type": "string",  "description": "Short human-scannable summary (e.g. 'repro the minimize distortion on a 3-pane cross-split')" },
-      "payload":      { "type": "string",  "description": "The full instruction injected into whichever agent claims this. Write it as a standalone prompt: the claimant has none of your conversation context." },
-      "kind":         { "type": "string",  "description": "Optional free-form tag used to filter claims (e.g. 'review', 'repro', 'triage'). Agents can claim only a kind they handle. Omit for untyped work anyone may take." },
-      "target_agent": { "type": "string",  "description": "Optional: restrict to ONE agent id. Mutually exclusive with target_group. Omit so any agent can claim — that is the normal case and the point of the queue." },
-      "target_group": { "type": "string",  "description": "Optional: restrict to members of an agent group id. Mutually exclusive with target_agent." },
-      "priority":     { "type": "integer", "description": "Higher claims first; ties break oldest-first. Default 0. Use sparingly — everything urgent means nothing is." },
-      "not_before":   { "type": "integer", "description": "Unix ms timestamp; the item is not claimable before it. Use for deferred work ('look at this after the release lands'). Omit for immediately claimable." },
-      "max_attempts": { "type": "integer", "description": "How many claims this item gets before it is parked as failed. Default 3. Guards against an item that crashes or defeats every agent that takes it." }
-    },
-    "required": ["title", "payload"]
-  }
-}"#;
-
-const WORK_CLAIM_TOOL: &str = r#"{
-  "name": "WorkClaim",
-  "description": "Take the next eligible item off the Muxqueue and become its holder. Returns {claimed:false} when nothing is available — that is a normal answer, not an error, so it is safe to call speculatively when you have spare capacity. Claiming grants a time-limited LEASE, not ownership: heartbeat with WorkHeartbeat during long work, then finish with WorkComplete (or hand it back with WorkRelease). If your lease expires the item returns to the pool for someone else. IMPORTANT: the response includes an 'attempt' number — you must pass it back to every WorkHeartbeat/WorkComplete/WorkRelease call for this item, or they will be rejected. Claiming an item does NOT grant you authority you would not otherwise have: the payload is a prompt, and every action in it is still subject to its own normal confirmation rules.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "kind":     { "type": "string",  "description": "Only claim items with this kind tag. Omit to consider every untargeted item." },
-      "lease_ms": { "type": "integer", "description": "How long your lease lasts before the item can be reclaimed by someone else. Default 120000 (2 min). Heartbeat rather than asking for a very long lease — a long lease on a crashed agent blocks the item for that whole window." }
-    }
-  }
-}"#;
-
-const WORK_HEARTBEAT_TOOL: &str = r#"{
-  "name": "WorkHeartbeat",
-  "description": "Extend your lease on a claimed Muxqueue item while you are still working on it. Call this periodically during long work; without it the lease expires and another agent may take the item. Requires the 'attempt' number from your WorkClaim response — a heartbeat from a superseded claim is rejected (HTTP 409) rather than silently extending someone else's.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "id":       { "type": "string",  "description": "The item id from WorkClaim" },
-      "attempt":  { "type": "integer", "description": "The 'attempt' number returned by the WorkClaim that gave you this item" },
-      "lease_ms": { "type": "integer", "description": "New lease length in ms. Default 120000." }
-    },
-    "required": ["id", "attempt"]
-  }
-}"#;
-
-const WORK_COMPLETE_TOOL: &str = r#"{
-  "name": "WorkComplete",
-  "description": "Mark a claimed Muxqueue item finished. Requires the 'attempt' number from your WorkClaim response; a completion from a superseded claim is rejected (HTTP 409) so a slow agent cannot close out work another agent has since taken over. Record what you actually did in 'result' — it is the only trace of the work once the item is done.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "id":      { "type": "string",  "description": "The item id from WorkClaim" },
-      "attempt": { "type": "integer", "description": "The 'attempt' number returned by the WorkClaim that gave you this item" },
-      "result":  { "type": "string",  "description": "REQUIRED. What was done, or what the outcome was. Include links (PR, issue) where relevant. Once an item is done this is the only record that it happened — a completion with no result silently destroys the trace." }
-    },
-    "required": ["id", "attempt", "result"]
-  }
-}"#;
-
-const WORK_RELEASE_TOOL: &str = r#"{
-  "name": "WorkRelease",
-  "description": "Hand a claimed Muxqueue item back to the pool because you cannot do it — wrong capabilities, blocked on something, out of context budget. Prefer this over letting your lease silently expire: it frees the item immediately and records why. Note that a release still consumes one of the item's attempts, and releasing on its FINAL attempt parks it as failed rather than reopening it — an item nobody can do should stop circulating. Requires the 'attempt' number from your WorkClaim response.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "id":      { "type": "string",  "description": "The item id from WorkClaim" },
-      "attempt": { "type": "integer", "description": "The 'attempt' number returned by the WorkClaim that gave you this item" },
-      "reason":  { "type": "string",  "description": "Why you are handing it back. This is what the next claimant (or a human) sees." }
-    },
-    "required": ["id", "attempt"]
-  }
-}"#;
-
-const WORK_LIST_TOOL: &str = r#"{
-  "name": "WorkList",
-  "description": "List Muxqueue items — the shared backlog across every agent on this machine. Use it to see what is outstanding before enqueueing something (avoid duplicates), to check on work you enqueued, or to find out who is currently holding what. Read-only.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "state": { "type": "string", "description": "Filter by state: open, claimed, done, failed, cancelled. Omit for all states." },
-      "limit": { "type": "integer", "description": "Max items to return (1-500). Default 50." }
-    }
-  }
-}"#;
-
-const CRON_CREATE_TOOL: &str = r#"{
-  "name": "CronCreate",
-  "description": "AgentMux's own cross-agent cron — NOT the same tool as the native (non-mcp__agentmux__-prefixed) CronCreate your harness may also expose, which schedules only your own session's future turn. Use this one specifically to target a DIFFERENT agent (or when you need AgentMux-persisted delivery independent of any single session). Creates a persistent scheduled cron job that survives agent pane restarts. Fires the prompt on a UTC cron schedule by injecting it into the target agent. Unlike Loop, cron jobs persist as long as agentmux-srv is running. Returns a job id and the next scheduled fire time. If you're scheduling your OWN future turn instead, prefer the native CronCreate/ScheduleWakeup tools — no cross-agent envelope overhead.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "name":       { "type": "string",  "description": "Human-readable label for the job (e.g. 'daily-standup-check')" },
-      "expression": { "type": "string",  "description": "5-field UTC cron expression: 'min hour dom month dow' (e.g. '0 9 * * 1-5' = 9am weekdays). Standard cron syntax; ranges, lists, and step values are supported." },
-      "prompt":     { "type": "string",  "description": "The prompt or slash command to inject at each scheduled fire" },
-      "to":         { "type": "string",  "description": "Target agent id to inject into. Required." },
-      "max_fires":  { "type": "integer", "description": "Auto-disable after this many fires (the job row stays in DB for audit; use CronDelete to remove it). Omit for unlimited." },
-      "max_age_secs": { "type": "integer", "description": "Auto-disable this many seconds after creation, regardless of fire count (a hard staleness/stuck-loop bound, matching the spirit of native CronCreate's 7-day auto-expiry). Omit for no expiry — appropriate for genuinely long-running cross-agent automations; set this when babysitting something that should have a natural end (e.g. 'stop checking this PR after 6 hours even if it's still open')." }
-    },
-    "required": ["name", "expression", "prompt", "to"]
-  }
-}"#;
-
-const CRON_DELETE_TOOL: &str = r#"{
-  "name": "CronDelete",
-  "description": "Delete a persistent AgentMux cross-agent cron job (created via this same mcp__agentmux__ tool family's CronCreate, not the native per-session one) by id. Stops all future fires immediately.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "id": { "type": "string", "description": "The job id returned by CronCreate or CronList" }
-    },
-    "required": ["id"]
-  }
-}"#;
-
-const CRON_LIST_TOOL: &str = r#"{
-  "name": "CronList",
-  "description": "List all persistent AgentMux cross-agent cron jobs (created via this same mcp__agentmux__ tool family, not the native per-session ones). Returns each job's id, name, expression, next fire time, fire count, and enabled state.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-const CRON_PAUSE_TOOL: &str = r#"{
-  "name": "CronPause",
-  "description": "Pause a persistent AgentMux cross-agent cron job (this mcp__agentmux__ tool family, not the native per-session one — native cron has no pause/resume, only delete). The job definition is kept in the DB but no fires occur until CronResume is called.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "id": { "type": "string", "description": "The job id to pause" }
-    },
-    "required": ["id"]
-  }
-}"#;
-
-const CRON_RESUME_TOOL: &str = r#"{
-  "name": "CronResume",
-  "description": "Resume a paused AgentMux cross-agent cron job (this mcp__agentmux__ tool family). The job will fire at its next scheduled UTC time.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "id": { "type": "string", "description": "The job id to resume" }
-    },
-    "required": ["id"]
-  }
-}"#;
-
-const MEMORY_LIST_TOOL: &str = r#"{
-  "name": "MemoryList",
-  "description": "List your own native memory (brain) markdown files. Returns each file's filename, whether it is the index, its metadata_type, size in bytes, and last-modified time. Use it to see what you've remembered before reading or writing a specific file. Takes no arguments.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-const MEMORY_READ_TOOL: &str = r#"{
-  "name": "MemoryRead",
-  "description": "Read one of your own native memory (brain) markdown files by filename. Returns its content. Get valid filenames from MemoryList.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "filename": { "type": "string", "description": "The memory file to read (from MemoryList)" }
-    },
-    "required": ["filename"]
-  }
-}"#;
-
-const MEMORY_WRITE_TOOL: &str = r#"{
-  "name": "MemoryWrite",
-  "description": "Create or overwrite one of your own native memory (brain) markdown files. The write is atomic. Use it to persist notes/context for your future self across conversations. Every write is retained as a version (see MemoryHistory) — nothing is ever silently lost.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "filename": { "type": "string", "description": "The memory file to write (created if absent, overwritten if present)" },
-      "content":  { "type": "string", "description": "Full markdown content to store in the file" },
-      "provenance": {
-        "type": "object",
-        "description": "Optional context for why you're writing this — helps a human reviewing history later. Omit for an ordinary write from your own reasoning.",
-        "properties": {
-          "source": { "type": "string", "description": "\"human\" if directly instructed by the operator, \"jekt\" if this write is a direct response to jekt content still in your context, omit otherwise (defaults to agent_inferred)" },
-          "detail":  { "type": "object", "description": "Extra structured context — e.g. the jekt's marker fields (FROM/TIER/TRUST/DELIVERY/MSGID) when source is \"jekt\"" }
-        },
-        "required": ["source"]
-      }
-    },
-    "required": ["filename", "content"]
-  }
-}"#;
-
-const MEMORY_HISTORY_TOOL: &str = r#"{
-  "name": "MemoryHistory",
-  "description": "List every recorded version of one of your own native memory (brain) markdown files, newest first. Each entry shows who/what wrote it (source: human, agent_inferred, jekt, external_fs_write, or revert) and when. Use it to review how a memory file changed over time, or to find a version id to pass to MemoryDiff/MemoryRevert.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "filename": { "type": "string", "description": "The memory file to show history for (from MemoryList)" }
-    },
-    "required": ["filename"]
-  }
-}"#;
-
-const MEMORY_DIFF_TOOL: &str = r#"{
-  "name": "MemoryDiff",
-  "description": "Show a line-based diff between two recorded versions of a memory file. Get version ids from MemoryHistory.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "from_version_id": { "type": "string", "description": "The earlier version id (from MemoryHistory)" },
-      "to_version_id":   { "type": "string", "description": "The later version id (from MemoryHistory)" }
-    },
-    "required": ["from_version_id", "to_version_id"]
-  }
-}"#;
-
-const MEMORY_REVERT_TOOL: &str = r#"{
-  "name": "MemoryRevert",
-  "description": "Restore a memory file's live content to a prior recorded version. This does NOT delete history — it records a new version (source: \"revert\") whose content matches the target, same as `git revert`. Use it to undo a bad or fabricated memory write once you've confirmed via MemoryHistory/MemoryDiff which version to restore.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "filename": { "type": "string", "description": "The memory file to revert" },
-      "target_version_id": { "type": "string", "description": "The version id to restore (from MemoryHistory)" }
-    },
-    "required": ["filename", "target_version_id"]
-  }
-}"#;
-
-const PRESET_LIST_TOOL: &str = r#"{
-  "name": "PresetList",
-  "description": "List the presets available to you (summary fields only). A preset is a provider-agnostic config bundle — instructions, context files, MCP servers, and skills. Use it to discover presets before fetching one in full with PresetGet. Takes no arguments.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-const PRESET_GET_TOOL: &str = r#"{
-  "name": "PresetGet",
-  "description": "Fetch a full preset object by id or name. With BOTH id and name omitted, returns your OWN bound preset (the one you are currently configured with). Use PresetList to discover ids/names.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "id":   { "type": "string", "description": "Preset id to fetch (optional)" },
-      "name": { "type": "string", "description": "Preset name to fetch (optional)" }
-    }
-  }
-}"#;
-
-const IDENTITY_ACCOUNTS_TOOL: &str = r#"{
-  "name": "IdentityAccounts",
-  "description": "List your own linked identity accounts. Returns each account's account_id, provider, name, kind, status, masked_tail, and updated_at. Secrets are never returned — only masked tails. Use it to see which provider accounts you can use and to get account_ids for IdentityValidate. Takes no arguments.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {}
-  }
-}"#;
-
-const IDENTITY_VALIDATE_TOOL: &str = r#"{
-  "name": "IdentityValidate",
-  "description": "Live-probe one of your own linked accounts against its provider using the stored key, to confirm the credential still works. You never supply a secret — pass an account_id from IdentityAccounts. Returns valid, status, masked_tail, and error.",
-  "inputSchema": {
-    "type": "object",
-    "properties": {
-      "account_id": { "type": "string", "description": "One of your linked accounts (from IdentityAccounts)" }
-    },
-    "required": ["account_id"]
-  }
-}"#;
+mod tool_schemas;
+use tool_schemas::*;
+mod window_capture;
+use window_capture::*;
 
 #[tokio::main]
 async fn main() {
@@ -961,797 +265,6 @@ fn require_agent_env(local_url: &str, auth_key: &str, block_id: &str) -> Result<
 /// local users. `agentmux-mcp` can't import `agentmux-srv`'s function
 /// directly (separate crate/process), so this replicates its exact logic
 /// instead of inventing a new convention.
-fn capture_window_dir() -> std::path::PathBuf {
-    let base = std::env::var("AGENTMUX_DATA_HOME")
-        .ok()
-        .filter(|d| !d.is_empty())
-        .map(std::path::PathBuf::from)
-        .unwrap_or_else(|| {
-            dirs::home_dir()
-                .unwrap_or_else(|| std::path::PathBuf::from("/"))
-                .join(".agentmux")
-        });
-    base.join("tmp/capture-window")
-}
-
-const CAPTURE_RETENTION: std::time::Duration = std::time::Duration::from_secs(60 * 60);
-
-/// Same bug class, same fix, as `agentmux-srv`'s `prune_old_screenshots`
-/// (`ui_handlers.rs`, reagent P2, PR #2662) — reapplied here per reagent's
-/// review of this tool's own PR (#2709 round 1), which found it wasn't
-/// reused. Best-effort, on the write path, PNG-only.
-fn prune_old_captures(dir: &std::path::Path) {
-    let Ok(entries) = std::fs::read_dir(dir) else {
-        return;
-    };
-    let now = std::time::SystemTime::now();
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.extension().and_then(|e| e.to_str()) != Some("png") {
-            continue;
-        }
-        let Ok(metadata) = entry.metadata() else { continue };
-        let Ok(modified) = metadata.modified() else { continue };
-        let Ok(age) = now.duration_since(modified) else { continue };
-        if age > CAPTURE_RETENTION {
-            let _ = std::fs::remove_file(&path);
-        }
-    }
-}
-
-/// Max ancestor hops to walk from this MCP process before giving up —
-/// bounds the fallback search (see `own_instance_pids`'s doc comment for why
-/// this is a fallback, not the primary signal), while never walking so far
-/// up the tree (e.g. to `explorer.exe`) that "same instance" degrades into
-/// "everything on the desktop."
-const OWN_INSTANCE_ANCESTOR_HOPS: usize = 8;
-
-/// PIDs `CaptureWindow` must treat as "this agent's own AgentMux instance"
-/// and always exclude — reagent P0 (PR #2709 round 3).
-///
-/// **Primary signal: `AGENTMUX_APP_PATH`.** Every process belonging to this
-/// exact running instance (portable build or install) is launched from
-/// under that one directory — confirmed directly (not assumed): this
-/// agent's own host process's `Process::exe()` path
-/// (`...\agentmux-0.55.18+....-x64-portable\runtime\agentmux-0.55.18.exe`)
-/// starts with `AGENTMUX_APP_PATH`
-/// (`...\agentmux-0.55.18+....-x64-portable`) exactly. This is more precise
-/// than matching on version number alone — it correctly tells apart two
-/// separate instances that happen to share a version (e.g. two portable
-/// builds of the same release), which a version-string match would
-/// conflate.
-///
-/// **Fallback signal: bounded process-ancestor walk.** Kept as a second,
-/// additive layer (union, not replacement) for the case `AGENTMUX_APP_PATH`
-/// isn't set (confirmed present for `AGENTMUX_RUNTIME_MODE=portable`; not
-/// separately confirmed for `task dev` instances). **This alone is not
-/// reliable** — verified directly by testing: Windows recycles PIDs and a
-/// process's recorded parent-PID can point at an already-exited process, so
-/// `sys.process(stale_ppid)` returns `None` and the walk silently stops
-/// short of the real ancestor chain. Confirmed this exact failure in
-/// testing (the walk never reached this instance's own host process). Kept
-/// only as defense-in-depth alongside the path-based signal, never alone.
-///
-/// Fails safe: an unresolvable `pid()` on a *candidate* window (checked by
-/// the caller, `CaptureWindow`'s handler) is treated as "assume it's mine,
-/// exclude it" — so a window this function's own signals can't positively
-/// place is still excluded, not silently let through.
-fn own_instance_pids() -> std::collections::HashSet<u32> {
-    let sys = sysinfo::System::new_all();
-    let mut result = std::collections::HashSet::new();
-
-    if let Ok(app_path) = std::env::var("AGENTMUX_APP_PATH") {
-        if !app_path.is_empty() {
-            let app_path = std::path::Path::new(&app_path);
-            for (pid, proc) in sys.processes() {
-                if proc.exe().map(|e| e.starts_with(app_path)).unwrap_or(false) {
-                    result.insert(pid.as_u32());
-                }
-            }
-        }
-    }
-
-    let my_pid = sysinfo::Pid::from(std::process::id() as usize);
-    let mut ancestors = vec![my_pid];
-    let mut current = my_pid;
-    for _ in 0..OWN_INSTANCE_ANCESTOR_HOPS {
-        let Some(proc) = sys.process(current) else { break };
-        let Some(parent) = proc.parent() else { break };
-        ancestors.push(parent);
-        current = parent;
-    }
-    result.extend(ancestors.iter().map(|p| p.as_u32()));
-    for (pid, proc) in sys.processes() {
-        let Some(ppid) = proc.parent() else { continue };
-        if !ancestors.contains(&ppid) {
-            continue;
-        }
-        if proc.name().to_string_lossy().to_lowercase().starts_with("agentmux") {
-            result.insert(pid.as_u32());
-        }
-    }
-    result
-}
-
-/// Capture tier for one target window — see
-/// `docs/specs/SPEC_AGENT_UNRESTRICTED_CAPTURE_WITH_ACCOUNTABILITY_2026_08_30.md`
-/// §3. Keyed on WHAT is being captured, never on who is asking: identity is
-/// already proven upstream (`sign_ui_automation_auth`), so the open question
-/// is what a proven identity may reach.
-///
-/// This replaces the previous binary `!is_self` exclusion. That rule came from
-/// `SPEC_AGENT_UI_AUTOMATION_CLICK_SCREENSHOT_2026_08_18.md` §6, which itself
-/// says cross-agent targeting *"if it's ever wanted at all"* should be a
-/// distinct capability — i.e. it defaulted closed because no mechanism existed
-/// to be selective, not because open had been judged wrong. The repo owner has
-/// since directed that agents be able to capture anything. This enum is that
-/// mechanism.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum CaptureTier {
-    /// T1 — another pane inside the caller's OWN instance. Its window hosts
-    /// every agent's pane, which is exactly what the old rule blocked.
-    SameInstance,
-    /// T2 — a different AgentMux instance owned by the same OS user.
-    OtherInstance,
-    /// T3 — a window owned by a DIFFERENT OS user. The one tier that crosses a
-    /// human rather than an agent boundary: that user never consented to this
-    /// app's agents, and no in-app notification can reach them.
-    OtherUser,
-    /// T4 — a non-AgentMux window owned by the same OS user (their browser,
-    /// their password manager). PR #2709 round 2 caught the original unscoped
-    /// tool capturing a KeePass window; allowed now, but always audited.
-    ForeignApp,
-}
-
-impl CaptureTier {
-    /// Phase 1 `allow` defaults (spec §3). Only T3 is closed, and only because
-    /// it is a human-to-human boundary — every agent-to-agent tier is open.
-    fn allowed(self) -> bool {
-        !matches!(self, CaptureTier::OtherUser)
-    }
-
-    fn label(self) -> &'static str {
-        match self {
-            CaptureTier::SameInstance => "T1-same-instance",
-            CaptureTier::OtherInstance => "T2-other-instance",
-            CaptureTier::OtherUser => "T3-other-user",
-            CaptureTier::ForeignApp => "T4-foreign-app",
-        }
-    }
-}
-
-/// The OS user id owning the calling process, used to place a target window on
-/// the near or far side of the T3 boundary.
-///
-/// Fails CLOSED, matching `own_instance_pids()`'s existing discipline: if this
-/// returns `None` every window resolves to `OtherUser` and capture is denied.
-/// An unresolvable owner is exactly the case where guessing "probably mine"
-/// would be how a cross-user capture slips through.
-fn current_user_id(sys: &sysinfo::System) -> Option<String> {
-    let me = sysinfo::get_current_pid().ok()?;
-    sys.process(me)
-        .and_then(|p| p.user_id())
-        .map(|u| u.to_string())
-}
-
-/// A completed capture. Richer than the bare message string it replaces so
-/// `audit_log_capture_window` can record WHAT was captured and at which tier,
-/// rather than only the caller's query and a success/failure flag.
-#[derive(Debug)]
-struct CaptureOutcome {
-    /// Human-facing tool result, returned to the caller verbatim.
-    message: String,
-    tier: CaptureTier,
-    /// Resolved target — `pid=N title="..."`.
-    target: String,
-    /// Hex SHA-256 of the saved PNG; `None` if hashing failed (best-effort).
-    image_sha256: Option<String>,
-}
-
-/// One top-level window visible to this process, with the metadata both
-/// `DiscoverWindows` and `CaptureWindow` need. Shared via
-/// `enumerate_agentmux_windows()` so the two tools can't drift on what counts
-/// as "AgentMux's own window", "the calling agent's own instance", or which
-/// tier a target falls in.
-struct AgentMuxWindowInfo {
-    window: xcap::Window,
-    pid: u32,
-    title: String,
-    exe_path: String,
-    is_self: bool,
-    /// Whether the owning process is an AgentMux process at all — T4 targets
-    /// are enumerated now, so this is no longer implied by presence in the list.
-    is_agentmux: bool,
-    tier: CaptureTier,
-}
-
-/// Enumerate every top-level OS window, AgentMux-owned or not, each tagged
-/// with the `CaptureTier` that decides whether it may be captured.
-///
-/// `is_agentmux` is matched via `app_name()`, not `title()` — AgentMux's own
-/// process names are version-stamped, e.g. `agentmux-0.55.18`, not a single
-/// fixed string, but they all share the `agentmux` prefix (reagent P1, PR
-/// #2709 round 2). That flag no longer decides *inclusion* — non-AgentMux
-/// windows are T4 capture targets — but it still gates disclosure: they stay
-/// out of `DiscoverWindows`' default listing and their titles are withheld
-/// from candidate/miss lists (`candidate_label`), which is what keeps round
-/// 2's KeePass-title leak closed now that they're enumerated at all.
-///
-/// `is_self` marks windows belonging to the calling agent's OWN instance
-/// (`own_instance_pids()`).
-///
-/// **Historical note (reagent P2 on PR #2845):** this comment previously said
-/// `is_self` windows were "hard-excluded downstream by `capture_window_impl`
-/// … the actual isolation boundary, not a nicety" (reagent P0, PR #2709
-/// round 3). That is no longer true and the reasoning has been superseded, not
-/// merely relaxed: the boundary it protected —
-/// `SPEC_AGENT_UI_AUTOMATION_CLICK_SCREENSHOT_2026_08_18.md` §6's own-pane-only
-/// default — was an unratified recommendation that defaulted closed because no
-/// mechanism existed to be selective. `is_self` now resolves to
-/// `CaptureTier::SameInstance`, which is allowed. See
-/// `SPEC_AGENT_UNRESTRICTED_CAPTURE_WITH_ACCOUNTABILITY_2026_08_30.md`.
-fn enumerate_agentmux_windows() -> Result<Vec<AgentMuxWindowInfo>> {
-    let windows =
-        xcap::Window::all().map_err(|e| anyhow::anyhow!("failed to enumerate windows: {e}"))?;
-    let own_pids = own_instance_pids();
-    let sys = sysinfo::System::new_all();
-    let me = current_user_id(&sys);
-    let mut out = Vec::new();
-    for window in windows {
-        let is_agentmux = window
-            .app_name()
-            .map(|a| a.to_lowercase().starts_with("agentmux"))
-            .unwrap_or(false);
-        // Non-AgentMux windows are no longer skipped — they are T4 targets.
-        // They stay out of `DiscoverWindows`' default listing (see its
-        // `include_foreign` arg) so ordinary discovery doesn't disclose the
-        // titles of a user's unrelated applications as a side effect.
-        let pid = window.pid().unwrap_or(0);
-        // Fails safe: an unresolvable pid is treated as "assume it's mine"
-        // rather than silently dropped, so `DiscoverWindows` still surfaces
-        // that the window exists instead of hiding it.
-        //
-        // That promise is real again as of reagentx P2 on PR #2845. Briefly it
-        // wasn't: an unresolvable OWNER resolves the window to T3, and
-        // `DiscoverWindows` was dropping non-allowed tiers outright, so such a
-        // window vanished from the listing even with `include_self`. It is now
-        // listed with `title`/`exe_path` redacted, which keeps the disclosure
-        // closed without reintroducing the hiding this comment warns against.
-        let is_self = pid == 0 || own_pids.contains(&pid);
-        let proc = sys.process(sysinfo::Pid::from(pid as usize));
-        let title = window.title().unwrap_or_default();
-        let exe_path = proc
-            .and_then(|p| p.exe())
-            .map(|p| p.display().to_string())
-            .unwrap_or_default();
-
-        // Tier resolution. The OS-user check runs FIRST and outranks
-        // everything else: a window owned by another human is T3 no matter
-        // what process owns it, including another AgentMux instance of theirs.
-        let owner = proc.and_then(|p| p.user_id()).map(|u| u.to_string());
-        let tier = match (&me, &owner) {
-            // Both known and different — the human boundary.
-            (Some(mine), Some(theirs)) if mine != theirs => CaptureTier::OtherUser,
-            // Either side unresolvable — fail closed, same discipline as
-            // `is_self` above. Guessing "probably mine" here is precisely how
-            // a cross-user capture would slip through.
-            (None, _) | (_, None) => CaptureTier::OtherUser,
-            _ if is_self => CaptureTier::SameInstance,
-            _ if is_agentmux => CaptureTier::OtherInstance,
-            _ => CaptureTier::ForeignApp,
-        };
-
-        out.push(AgentMuxWindowInfo {
-            window,
-            pid,
-            title,
-            exe_path,
-            is_self,
-            is_agentmux,
-            tier,
-        });
-    }
-    Ok(out)
-}
-
-/// Cheap heuristic for "this capture is probably a blank/unpainted frame,
-/// not a real render" — sample ~200 evenly-spaced pixels and check they're
-/// all within a small tolerance of the first one. Not real image analysis
-/// (a legitimately solid-color themed window would also trip this) —
-/// deliberately cheap and approximate, used only to decide whether a short
-/// bounded retry is worth attempting, and to set `likely_unrendered` as a
-/// hint, never as a hard failure.
-fn looks_unrendered(img: &image::RgbaImage) -> bool {
-    let pixels = img.as_raw();
-    if pixels.len() < 4 {
-        return true;
-    }
-    let (r0, g0, b0) = (pixels[0], pixels[1], pixels[2]);
-    let pixel_count = pixels.len() / 4;
-    let step = (pixel_count / 200).max(1);
-    let mut sampled = 0usize;
-    for i in (0..pixel_count).step_by(step) {
-        let idx = i * 4;
-        sampled += 1;
-        if pixels[idx].abs_diff(r0) > 8 || pixels[idx + 1].abs_diff(g0) > 8 || pixels[idx + 2].abs_diff(b0) > 8 {
-            return false;
-        }
-    }
-    sampled > 0
-}
-
-/// Label for a window in the AUDIT TRAIL.
-///
-/// Fuller than `candidate_label` where it can be — reagentx P2 on PR #2845
-/// correctly spotted that reusing the caller-facing label lost a T4 window's
-/// title on a post-resolution failure, giving foreign-app failures less audit
-/// detail than AgentMux ones for no reason.
-///
-/// But NOT unconditionally full, because that finding's premise — "the audit
-/// is server-side only" — does not hold here. The trail is an NDJSON file in
-/// this instance's own data dir, and agents have shell access; it was read
-/// with `tail` during this PR's own verification. It is therefore not a
-/// confidential sink.
-///
-/// So audit detail follows the ALLOW decision, not AgentMux-ness:
-///   - allowed tier → full title. The agent could capture the window and read
-///     the title off the pixels anyway, so recording it discloses nothing new.
-///   - withheld tier → pid and tier only. Recording a T3 title would hand the
-///     agent, via a file it can read, exactly the content the tier denied it —
-///     reopening through the log the same oracle closed in the title-miss path.
-fn audit_target_label(w: &AgentMuxWindowInfo) -> String {
-    if w.tier.allowed() {
-        format!("pid={} title={:?}", w.pid, w.title)
-    } else {
-        format!("pid={} {} <title withheld>", w.pid, w.tier.label())
-    }
-}
-
-/// Label for a window in a CALLER-FACING disambiguation/candidate list.
-///
-/// Stricter than `audit_target_label` above, because the audiences differ: this
-/// one is returned to the agent, so it withholds the TITLE of every
-/// non-AgentMux window — reagent P1 on PR #2845. Foreign windows became
-/// capturable with the tier model, so any list built from the capturable set
-/// would otherwise disclose a user's unrelated app titles (their browser,
-/// their password manager) as a side effect of a miss or an ambiguous match,
-/// bypassing `DiscoverWindows`' own `include_foreign` opt-in. The pid alone
-/// disambiguates, and is what the caller needs anyway.
-fn candidate_label(w: &AgentMuxWindowInfo) -> String {
-    if w.is_agentmux {
-        format!("pid={} title={:?}", w.pid, w.title)
-    } else {
-        format!("pid={} <non-AgentMux window; pass include_foreign to DiscoverWindows to identify it>", w.pid)
-    }
-}
-
-/// The actual `CaptureWindow` logic — window enumeration, own-instance and
-/// third-party-app exclusion, targeting (by pid or by title), capture, and
-/// save. Extracted to its own function (rather than living inline in the
-/// `"CaptureWindow" =>` match arm) so its `Result` can be captured once and
-/// unconditionally audit-logged by the caller before propagating — see
-/// `audit_log_capture_window`.
-///
-/// `index: None` with more than one `title_contains` match is an error
-/// listing every candidate, NOT a silent pick of match 0 — see
-/// docs/reports/REPORT_AGENT_SCREENSHOT_WINDOW_CONTROL_BLOCKERS_2026_08_24.md
-/// §1 for the real incident (an ambiguous match silently captured an
-/// unrelated, sensitive window) this specifically fixes. `index: Some(i)`
-/// is still honored directly against whatever matched, unchanged from the
-/// original behavior, for callers who already disambiguate explicitly.
-fn capture_window_impl(
-    title_contains: Option<&str>,
-    index: Option<usize>,
-    pid: Option<u32>,
-    // Filled the moment a target is RESOLVED, so a failure after that point —
-    // a denied T3 pid, a failed capture_image, a failed PNG write — is still
-    // audited with its tier and target instead of nulls (codex P2 on PR #2845).
-    // Without it the log cannot tell "no such window" apart from "blocked
-    // cross-user attempt", which is the entry a reviewer most wants to find.
-    resolved: &mut Option<(CaptureTier, String)>,
-) -> Result<CaptureOutcome> {
-    let all = enumerate_agentmux_windows()?;
-    // Tier gate replaces the old `!w.is_self` exclusion — see `CaptureTier`.
-    // Only T3 (a different OS user's window) is withheld; every agent-to-agent
-    // tier, including this instance's own window, is reachable.
-    let foreign: Vec<&AgentMuxWindowInfo> = all.iter().filter(|w| w.tier.allowed()).collect();
-
-    let target: &AgentMuxWindowInfo = if let Some(target_pid) = pid {
-        // A single host process can own multiple top-level windows sharing
-        // the same pid (agentmux-cef/src/browser_pane/hwnd.rs:200-204,
-        // agentmux-cef/src/commands/window/lifecycle.rs:410-426) — codex P1
-        // on PR #2810: the original `.find()` here silently captured
-        // whichever matching window enumerated first, which could be a
-        // pool/sub-window rather than the one the caller meant, reintroducing
-        // the wrong-window capture this PR exists to prevent. Same
-        // ambiguity-rejection shape as the title_contains branch below:
-        // `index` disambiguates, an unqualified ambiguous match does not.
-        let matches: Vec<&AgentMuxWindowInfo> = foreign
-            .iter()
-            .filter(|w| w.pid == target_pid)
-            .copied()
-            .collect();
-        if matches.is_empty() {
-            // `foreign` is already tier-filtered, so a withheld T3 window is
-            // absent from it and would otherwise be indistinguishable from a
-            // pid that doesn't exist — reported identically to the caller AND
-            // audited as `tier: null`. Look it up in the UNFILTERED set to
-            // separate "withheld" from "absent" (codex P2 / reagentx P2).
-            if let Some(blocked) = all.iter().find(|w| w.pid == target_pid) {
-                *resolved = Some((blocked.tier, audit_target_label(blocked)));
-                anyhow::bail!(
-                    "window pid={target_pid} is {} — capture is withheld for windows owned \
-                     by a different OS user; every other tier is available",
-                    blocked.tier.label()
-                );
-            }
-            anyhow::bail!(
-                "no window found for pid {target_pid} — call DiscoverWindows to see \
-                 current candidates"
-            );
-        }
-        match index {
-            Some(i) => *matches.get(i).ok_or_else(|| {
-                anyhow::anyhow!(
-                    "index {i} out of range — only {} window(s) matched pid {target_pid}",
-                    matches.len()
-                )
-            })?,
-            None if matches.len() == 1 => matches[0],
-            None => {
-                let candidates: Vec<String> = matches
-                    .iter()
-                    .map(|w| candidate_label(w))
-                    .collect();
-                anyhow::bail!(
-                    "{} windows matched pid {target_pid} — a single process can own \
-                     multiple top-level windows; pass an explicit index (0-based) \
-                     alongside pid to disambiguate. Candidates: {}",
-                    matches.len(),
-                    candidates.join("; ")
-                );
-            }
-        }
-    } else {
-        let title_contains = title_contains
-            .ok_or_else(|| anyhow::anyhow!("must provide either pid or title_contains"))?;
-        let needle = title_contains.to_lowercase();
-        let matches: Vec<&AgentMuxWindowInfo> = foreign
-            .iter()
-            .filter(|w| w.title.to_lowercase().contains(&needle))
-            .copied()
-            .collect();
-
-        if matches.is_empty() {
-            // A title matching ONLY a withheld window is recorded for the
-            // AUDIT — and deliberately changes nothing the caller sees.
-            //
-            // reagentx P1 on PR #2845: an earlier version bailed here with a
-            // distinguishable "exists but is T3" message. Withholding the
-            // literal title wasn't enough — differing *responses* are an
-            // existence oracle. An agent can probe substrings (binary-search
-            // over characters) and reconstruct a cross-user window's title
-            // without ever being granted capture, defeating the very boundary
-            // this tier exists to hold. Introduced while fixing the audit gap
-            // below it: making withheld-vs-absent legible to the auditor made
-            // it legible to the attacker too.
-            //
-            // The split that resolves it: the AUDIT is server-side and may
-            // know; the CALLER may not. So this records `resolved` and then
-            // falls through to the identical generic-miss message below —
-            // structurally the same code path and the same bytes either way,
-            // rather than two branches a future edit could let drift apart.
-            //
-            // The pid branch above is deliberately NOT symmetric: it does name
-            // the tier. A caller must already hold the pid to ask, `Discover
-            // Windows` never hands out T3 pids, and process ownership is
-            // already enumerable by anything with shell access — so the pid's
-            // existence is not a secret this tool is keeping, while a window
-            // TITLE is exactly the content it is.
-            if let Some(blocked) = all
-                .iter()
-                .find(|w| !w.tier.allowed() && w.title.to_lowercase().contains(&needle))
-            {
-                *resolved = Some((blocked.tier, audit_target_label(blocked)));
-            }
-            // Only lists AGENTMUX windows' titles, never every window on the
-            // desktop — reagent P2 (PR #2709 round 2): the original version
-            // dumped every visible window's title on any miss, which let a
-            // caller enumerate arbitrary window titles (confirmed in testing:
-            // it leaked a password manager's document title) with no real
-            // match required at all.
-            //
-            // The `is_agentmux` filter is load-bearing again as of the tier
-            // model — reagent P1 on PR #2845. `foreign` now includes T4
-            // non-AgentMux windows (they became capturable), so listing it
-            // wholesale would have re-opened exactly that leak, and would have
-            // bypassed `DiscoverWindows`' own `include_foreign` opt-in in the
-            // process: a caller could enumerate the user's desktop by
-            // deliberately missing.
-            let titles: Vec<&str> = foreign
-                .iter()
-                .filter(|w| w.is_agentmux)
-                .map(|w| w.title.as_str())
-                .filter(|t| !t.is_empty())
-                .collect();
-            if titles.is_empty() {
-                anyhow::bail!(
-                    "no AgentMux window title contains {title_contains:?} — \
-                     no other AgentMux windows are currently open"
-                );
-            }
-            anyhow::bail!(
-                "no AgentMux window title contains {title_contains:?}. \
-                 Open AgentMux window titles: {}",
-                titles.join(", ")
-            );
-        }
-
-        match index {
-            Some(i) => *matches.get(i).ok_or_else(|| {
-                anyhow::anyhow!(
-                    "index {i} out of range — only {} window(s) matched {title_contains:?}",
-                    matches.len()
-                )
-            })?,
-            None if matches.len() == 1 => matches[0],
-            None => {
-                // The fix for blocker #1 in the report: an ambiguous
-                // match with no explicit index used to silently capture
-                // index 0 despite this tool's own docstring claiming it
-                // would list candidates instead. It now actually does.
-                let candidates: Vec<String> = matches
-                    .iter()
-                    .map(|w| candidate_label(w))
-                    .collect();
-                anyhow::bail!(
-                    "{} windows matched {title_contains:?} — pass an explicit index (0-based), \
-                     or better, one of these pids directly (preferred: stable, unlike title). \
-                     Candidates: {}",
-                    matches.len(),
-                    candidates.join("; ")
-                );
-            }
-        }
-    };
-
-    *resolved = Some((target.tier, audit_target_label(target)));
-
-    let title = target.title.clone();
-
-    // Short bounded retry for a freshly-created window that hasn't
-    // painted its first real frame yet — see looks_unrendered()'s doc
-    // comment. std::thread::sleep (not tokio::time::sleep): this fn is
-    // plain sync code called from the async dispatch path, and the
-    // bounded ~800ms worst case is a deliberate, request-scoped wait
-    // directly serving this call, not background work stealing the
-    // runtime from anything else.
-    let mut image = target
-        .window
-        .capture_image()
-        .map_err(|e| anyhow::anyhow!("capture failed: {e}"))?;
-    let mut likely_unrendered = looks_unrendered(&image);
-    let mut attempts = 1;
-    while likely_unrendered && attempts < 3 {
-        std::thread::sleep(Duration::from_millis(400));
-        image = target
-            .window
-            .capture_image()
-            .map_err(|e| anyhow::anyhow!("capture failed: {e}"))?;
-        likely_unrendered = looks_unrendered(&image);
-        attempts += 1;
-    }
-
-    let dir = capture_window_dir();
-    std::fs::create_dir_all(&dir)
-        .map_err(|e| anyhow::anyhow!("failed to create capture dir {}: {e}", dir.display()))?;
-    let path = dir.join(format!("{}.png", uuid::Uuid::new_v4()));
-    image
-        .save(&path)
-        .map_err(|e| anyhow::anyhow!("failed to save capture to {}: {e}", path.display()))?;
-    // Unbounded growth over repeated/looped calls otherwise — same bug
-    // class as UIScreenshot's own screenshot dir (reagent P2, PR #2662)
-    // and reagent's own review of this PR (round 1) pointing out that
-    // fix wasn't reused here. Same fix, same shape: best-effort, on the
-    // write path.
-    prune_old_captures(&dir);
-
-    // codex P2 on PR #2810: no process start time is collected or checked
-    // anywhere, so claiming "this window was created recently" was
-    // unsupported and misleading for a mature window that just happens to
-    // render a near-uniform frame (looks_unrendered's own doc comment
-    // already acknowledges that legitimate case). State only what was
-    // actually observed.
-    let hint = if likely_unrendered {
-        " (likely_unrendered: true — the captured frame looks solid or \
-           near-solid-color even after retrying, which can mean the window \
-           hasn't painted its first real frame yet, or that it genuinely \
-           looks like this; consider calling CaptureWindow again shortly if \
-           that's unexpected)"
-    } else {
-        ""
-    };
-    Ok(CaptureOutcome {
-        message: format!(
-            "Captured window {title:?} to {}{hint} — use Read on that path to view it yourself, or OpenMedia to show it to the user.",
-            path.display()
-        ),
-        tier: target.tier,
-        // The resolved target, not the query string — an audit reviewer needs
-        // to know WHAT was captured, which a `title_contains` substring does
-        // not tell them.
-        target: format!("pid={} title={:?}", target.pid, target.title),
-        image_sha256: sha256_file(&path),
-    })
-}
-
-/// Hex SHA-256 of a captured PNG, recorded in the audit trail so a leaked
-/// screenshot can be traced back to the call that produced it (spec §6).
-/// Best-effort — a hashing failure yields `None` and must never fail the
-/// capture, matching `audit_log_capture_window`'s own "never break the tool"
-/// discipline.
-fn sha256_file(path: &std::path::Path) -> Option<String> {
-    use sha2::{Digest, Sha256};
-    let bytes = std::fs::read(path).ok()?;
-    let mut hasher = Sha256::new();
-    hasher.update(&bytes);
-    Some(format!("{:x}", hasher.finalize()))
-}
-
-/// Audit trail for `CaptureWindow` — reagent P1 (PR #2709 round 4): the
-/// residual risk after rounds 2-3's scoping (a different AgentMux
-/// instance's window, which can belong to a different OS user on a shared
-/// machine) is a disclosure-across-a-human-boundary question, not a
-/// technical one a capability flag alone would fully answer — a real
-/// per-agent opt-in gate is tracked as a separate follow-up (no existing
-/// settings/enforcement mechanism to build it on today). Logging every
-/// call — who, what was requested, what happened — is the honest,
-/// shippable Phase-1 answer: best-effort (a logging failure must never
-/// break the tool itself), append-only NDJSON inside `capture_window_dir()`
-/// itself — `prune_old_captures`'s PNG-only extension filter already
-/// leaves a `.log` file in that same directory untouched, so this doesn't
-/// need (or want) its own separate directory alongside it.
-fn audit_log_capture_window(
-    query_desc: &str,
-    outcome: &Result<CaptureOutcome>,
-    // Resolved target/tier for the FAILURE paths — on success the outcome
-    // carries its own. See capture_window_impl's `resolved` param.
-    resolved: &Option<(CaptureTier, String)>,
-) {
-    let entry = serde_json::json!({
-        "timestamp": std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs())
-            .unwrap_or(0),
-        "agent_id": agent_slug().unwrap_or_else(|_| "unknown".to_string()),
-        "tool": "CaptureWindow",
-        // "pid=N" or "title_contains=\"...\"" — whichever targeting mode
-        // the caller used (SPEC_AGENT_APP_API_WINDOW_CONTROL_ROBUSTNESS_2026_08_24.md
-        // added pid-based targeting as an alternative to title matching).
-        "query": query_desc,
-        // The RESOLVED target and its tier, not just the query string — a
-        // reviewer needs to know what was actually captured, which a
-        // `title_contains` substring doesn't say. Absent on failure, since
-        // nothing was resolved (spec §6).
-        "tier": outcome
-            .as_ref()
-            .ok()
-            .map(|c| c.tier.label())
-            .or_else(|| resolved.as_ref().map(|(t, _)| t.label())),
-        "target": outcome
-            .as_ref()
-            .ok()
-            .map(|c| c.target.clone())
-            .or_else(|| resolved.as_ref().map(|(_, t)| t.clone())),
-        // Traces a leaked screenshot back to the call that produced it.
-        "image_sha256": outcome.as_ref().ok().and_then(|c| c.image_sha256.clone()),
-        // Phase 1 never redacts — recorded explicitly so an unredacted T2/T4
-        // capture is distinguishable at review time once Phase 3 lands, rather
-        // than the field's absence being ambiguous between "no" and "old entry".
-        "redacted": false,
-        "outcome": match outcome {
-            Ok(c) => serde_json::json!({"result": "success", "detail": c.message}),
-            Err(e) => serde_json::json!({"result": "error", "detail": e.to_string()}),
-        },
-    });
-    append_window_audit_log_entry(&entry);
-}
-
-/// One `DiscoverWindows` entry.
-///
-/// A withheld (T3) window is LISTED but redacted, not omitted — reagentx P2 on
-/// PR #2845 caught two contradictions the omit-entirely approach created:
-///   - the `is_self` fail-safe promised an unresolvable-pid window would
-///     "still surface ... instead of hiding it", which an unconditional drop
-///     broke (an unresolvable owner resolves to T3)
-///   - this tool advertises `capturable` so a caller can "see why a window is
-///     out of reach before trying", unreachable if every listed entry is
-///     capturable by construction
-///
-/// Redacting satisfies both while keeping the disclosure closed: `title` and
-/// `exe_path` (which embeds the owning OS username) are exactly what must not
-/// cross the human boundary, and they are the only fields dropped. A pid and
-/// "someone else owns this" are already available to anything with shell
-/// access, so surfacing them costs nothing — and it makes a wholesale
-/// user-id-resolution failure diagnosable instead of silently returning an
-/// empty list.
-fn window_listing_entry(w: &AgentMuxWindowInfo) -> Value {
-    if w.tier.allowed() {
-        json!({
-            "pid": w.pid,
-            "title": w.title,
-            "exe_path": w.exe_path,
-            "is_self": w.is_self,
-            "is_agentmux": w.is_agentmux,
-            "tier": w.tier.label(),
-            "capturable": true,
-        })
-    } else {
-        json!({
-            "pid": w.pid,
-            "title": null,
-            "exe_path": null,
-            "is_self": w.is_self,
-            "is_agentmux": w.is_agentmux,
-            "tier": w.tier.label(),
-            "capturable": false,
-            "withheld_reason": "owned by a different OS user; title and exe_path withheld",
-        })
-    }
-}
-
-/// Audit trail for `DiscoverWindows` — reagent P1 on PR #2810: this tool
-/// discloses `exe_path` (a full filesystem path, typically embedding the OS
-/// username) for windows belonging to OTHER AgentMux instances/users on a
-/// shared machine — the exact same disclosure-across-a-human-boundary risk
-/// `audit_log_capture_window` above already exists to log, but this tool
-/// shipped with none. Same shape, same file (a single window-related audit
-/// trail is easier to review than two): best-effort, never blocks the
-/// tool's own result.
-fn audit_log_discover_windows(include_self: bool, include_foreign: bool, windows: &[Value]) {
-    let entry = serde_json::json!({
-        "timestamp": std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs())
-            .unwrap_or(0),
-        "agent_id": agent_slug().unwrap_or_else(|_| "unknown".to_string()),
-        "tool": "DiscoverWindows",
-        // Both flags — reagentx P2 on PR #2845. `include_foreign` is the one
-        // that actually triggers this trail's reason for existing: it
-        // discloses non-AgentMux window titles and exe_paths. Recording only
-        // `include_self` left a reviewer scanning query strings unable to tell
-        // whether foreign disclosure was even requested, inferable only by
-        // picking through each entry's `is_agentmux` tag.
-        "query": format!("include_self={include_self} include_foreign={include_foreign}"),
-        "outcome": serde_json::json!({
-            "result": "success",
-            "window_count": windows.len(),
-            "windows": windows,
-        }),
-    });
-    append_window_audit_log_entry(&entry);
-}
-
-/// Shared append-only NDJSON writer for both window-tool audit trails above
-/// — best-effort (a logging failure must never break the tool itself),
-/// inside `capture_window_dir()` itself since `prune_old_captures`'s
-/// PNG-only extension filter already leaves a `.log` file in that same
-/// directory untouched, so this doesn't need (or want) its own directory.
-fn append_window_audit_log_entry(entry: &Value) {
-    let dir = capture_window_dir();
-    if std::fs::create_dir_all(&dir).is_err() {
-        return;
-    }
-    let line = format!("{entry}\n");
-    if let Ok(mut f) = std::fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(dir.join("capture-window-audit.log"))
-    {
-        use std::io::Write;
-        let _ = f.write_all(line.as_bytes());
-    }
-}
 
 /// The calling agent's slug (its `AGENTMUX_AGENT_ID`), injected by AgentMux into
 /// this MCP server's trusted environment. The App API identity/preset/memory
@@ -1759,7 +272,7 @@ fn append_window_audit_log_entry(entry: &Value) {
 /// output cannot reach those endpoints (no auth key in the PTY) nor override
 /// this value, so the slug cannot be forged. See
 /// SPEC_AGENT_APP_API_MCP_BINDINGS_2026_06_28.md §5.
-fn agent_slug() -> Result<String> {
+pub(crate) fn agent_slug() -> Result<String> {
     let slug = std::env::var("AGENTMUX_AGENT_ID").unwrap_or_default();
     if slug.is_empty() {
         anyhow::bail!(
